Implementación y readaptacion a nuevo modelo

+++ src/main/java/edu/fiuba/algo3/modelo/direccion/Abajo.java
@@ -7,7 +7,7 @@ import edu.fiuba.algo3.modelo.celda.Celda;
 public class Abajo implements Direccion {
     
     @Override
-    public Celda mover(Integer f, Integer c, ArrayList<Celda> esquinas) {
+    public Celda mover(long f, long c, ArrayList<Celda> esquinas) {
         for (Celda e : esquinas) {
             if (e.fila() == f + 1 && e.columna() == c) return e;
         }
