+++ src/main/java/edu/fiuba/algo3/modelo/direccion/Izquierda.java
@@ -6,7 +6,7 @@ import edu.fiuba.algo3.modelo.celda.Celda;
 public class Izquierda implements Direccion {
     
     @Override
-    public Celda mover(Integer f, Integer c, ArrayList<Celda> esquinas) {
+    public Celda mover(long f, long c, ArrayList<Celda> esquinas) {
         for (Celda e : esquinas) {
             if (e.fila() == f && e.columna() == c - 1) return e;
         }
